+++ squidb-addons/squidb-json/squidb-json-annotations/src/com/yahoo/squidb/json/JSONProperty.java
@@ -3,12 +3,12 @@
  * Copyrights licensed under the Apache 2.0 License.
  * See the accompanying LICENSE file for terms.
  */
-package com.yahoo.squidb.jackson;
+package com.yahoo.squidb.json;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Target;
 
 @Target(ElementType.FIELD)
-public @interface JacksonProperty {
+public @interface JSONProperty {
 
 }
