+++ squidb-addons/squidb-json/squidb-json-compiler/src/com/yahoo/squidb/json/JacksonPropertyGenerator.java
@@ -3,7 +3,7 @@
  * Copyrights licensed under the Apache 2.0 License.
  * See the accompanying LICENSE file for terms.
  */
-package com.yahoo.squidb.jackson;
+package com.yahoo.squidb.json;
 
 import com.yahoo.aptutils.model.DeclaredTypeName;
 import com.yahoo.aptutils.utils.AptUtils;
